+++ packages/dart/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog (Dart)
 
-## 1.1.6 (2025--)
+## 1.1.6 (2025-10-15)
 
 - `isEmail`: add `onlyLowerCase` parameter
 - Add `console` method